Move randomQuote redux state to quote state property
